move startConversations into utils
